CSP-566-Refactoring namespaces
