+++ src/SFA.DAS.AANHub.Application.UnitTests/Partners/Commands/CreatePartnerMemberCommandTests.cs
@@ -1,7 +1,7 @@
 ﻿using FluentAssertions;
 using NUnit.Framework;
 using SFA.DAS.AANHub.Api.Models;
-using SFA.DAS.AANHub.Application.Partners;
+using SFA.DAS.AANHub.Application.Partners.Commands;
 using SFA.DAS.AANHub.Domain.Entities;
 
 namespace SFA.DAS.AANHub.Application.UnitTests.Partners.Commands
